Move interactions with GitHub to a task queue
This uses the Cloud Tasks API to enqueue tasks, and a POST endpoint on
the same service (and version) to run the tasks. The tasks aren't CPU
intensive, but they're fallible and will be retried when they fail.

Because the /api/task endpoint is exposed and the
X-Appengine-Inbound-Appid header isn't set for Node.js on AppEngine, the
webhook approach of signing with a shared secret is used, except now the
secret is internal to the project.